recircle-cardscan-backend/requirements.txt
# FastAPI Core
fastapi==0.104.0
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Pydantic (Data Validation)
pydantic==2.5.0
pydantic-settings==2.0.0

# Environment Variables
python-dotenv==1.0.0

# PDF Processing
pdf2image==1.16.3
PyPDF2==3.0.1
poppler-utils==0.1.0

# Image Processing
Pillow==10.1.0

# OCR - Tesseract (Local, Free)
pytesseract==0.3.10

# Async File Operations
aiofiles==23.2.1

# HTTP Client (for testing)
httpx==0.25.0

# CORS Support (already in FastAPI but explicit)
starlette==0.27.0

# Type Hints
typing-extensions==4.8.0

# Google Gemini AI
google-generativeai==0.8.3

# Optional: AWS Textract (if you want to use instead of Tesseract)
# boto3==1.29.0

# Optional: Azure Computer Vision (if you want to use instead of Tesseract)
# azure-cognitiveservices-vision-computervision==0.9.0